requirement.txt
Flask
numpy
joblib
mysql-connector-python
